Memoize users list interest groups from Mailchimp
This call seems to be taking up to 30s to come back, which is sort of bonkers. I’m hopeful that memoizing it will help it not be quite so insane, though it means that we’ll have to shut down the reader before making interest group changes if we want to do so. Probably not a big deal, but worth noting.