Get more categories per fetch
This is a total hack. The worker was crashing because it didn’t see a category in the local cache from Mailchimp that it was getting a request for, so it was attempting to create that category and refresh its cache, but the category already did exist and the reason it wasn’t showing in the cache was because the pagination limit was smaller than the number of categories, so the creation would bomb out.

This does periodically time out, but it works enough to bandaid the solution until it can be properly paginated.